pyproject.toml
[project]
name = "etsy-shipping-profile-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "pycountry>=24.6.1",
    "pydantic-settings>=2.11.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pandas-stubs>=2.3.2.250926",
    "ruff>=0.13.1",
    "tqdm-stubs>=0.2.1",
]
